Periodically restart the scgi handler.
Add a `bin/restart.sh` script that can be easily called from a cron job
to restart the scgi handler, in order to "fix" the problem where the
handler eventually loses the ability to connect to the database.

Add a section to the end of INSTALL.markdown about setting up said cron
job.